Fixed an issue with incorrect left alignment
There's an issue with incorrect left alignment of the placeholder in
certain conditions.
You can see an example of the issue at: http://codepen.io/anon/pen/xdFib
In the example notice that the placeholder text is to the left of the
input field.

The fix was to add compensation for x difference like you have for the y
difference.
You can see the result of applying the fix to the example at:
http://codepen.io/anon/pen/numxC
The placeholder text is now correctly placed in the the input field.